:recycle: Extract buttons
