+++ lib/domain/record/record_page.dart
@@ -4,13 +4,14 @@ import 'package:pilll/components/organisms/pill/pill_mark.dart';
 import 'package:pilll/components/organisms/pill/pill_sheet.dart';
 import 'package:pilll/domain/initial_setting/migrate_info.dart';
 import 'package:pilll/domain/premium_trial/premium_trial_complete_modal.dart';
+import 'package:pilll/domain/record/components/button/record_page_button.dart';
 import 'package:pilll/domain/record/components/notification_bar/notification_bar.dart';
 import 'package:pilll/domain/record/components/notification_bar/notification_bar_store_parameter.dart';
 import 'package:pilll/domain/record/record_page_state.dart';
 import 'package:pilll/domain/record/record_page_store.dart';
 import 'package:pilll/domain/record/record_taken_information.dart';
-import 'package:pilll/domain/modal/release_note.dart';
 import 'package:pilll/domain/premium_trial/premium_trial_modal.dart';
+import 'package:pilll/domain/record/util/take.dart';
 import 'package:pilll/entity/pill_sheet.dart';
 import 'package:pilll/entity/pill_sheet_type.dart';
 import 'package:pilll/entity/setting.dart';
@@ -19,20 +20,16 @@ import 'package:pilll/error/error_alert.dart';
 import 'package:pilll/error/universal_error_page.dart';
 import 'package:pilll/error_log.dart';
 import 'package:pilll/service/pill_sheet.dart';
-import 'package:pilll/components/atoms/buttons.dart';
 import 'package:pilll/components/atoms/color.dart';
 import 'package:pilll/components/atoms/font.dart';
 import 'package:pilll/components/atoms/text_color.dart';
 import 'package:pilll/util/datetime/day.dart';
-import 'package:pilll/util/shared_preference/keys.dart';
 import 'package:pilll/util/toolbar/picker_toolbar.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:flutter_svg/svg.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:in_app_review/in_app_review.dart';
-import 'package:shared_preferences/shared_preferences.dart';
 
 class RecordPage extends HookWidget {
   @override
@@ -169,82 +166,10 @@ class RecordPage extends HookWidget {
         if (currentPillSheet != null)
           Positioned(
             bottom: 20,
-            child: _button(context, currentPillSheet, store),
+            child: RecordPageButton(currentPillSheet: currentPillSheet),
           ),
       ],
     );
-  }
-
-  Widget _button(
-      BuildContext context, PillSheet currentPillSheet, RecordPageStore store) {
-    if (currentPillSheet.allTaken)
-      return _cancelTakeButton(currentPillSheet, store);
-    else
-      return _takenButton(context, currentPillSheet, store);
-  }
-
-  Widget _takenButton(
-    BuildContext context,
-    PillSheet pillSheet,
-    RecordPageStore store,
-  ) {
-    return PrimaryButton(
-      text: "飲んだ",
-      onPressed: () async {
-        if (pillSheet.todayPillNumber == 1)
-          analytics.logEvent(name: "user_taken_first_day_pill");
-        analytics.logEvent(name: "taken_button_pressed", parameters: {
-          "last_taken_pill_number": pillSheet.lastTakenPillNumber,
-          "today_pill_number": pillSheet.todayPillNumber,
-        });
-        await _take(context, pillSheet, now(), store);
-      },
-    );
-  }
-
-  Widget _cancelTakeButton(PillSheet pillSheet, RecordPageStore store) {
-    return TertiaryButton(
-      text: "飲んでない",
-      onPressed: () {
-        analytics.logEvent(name: "cancel_taken_button_pressed", parameters: {
-          "last_taken_pill_number": pillSheet.lastTakenPillNumber,
-          "today_pill_number": pillSheet.todayPillNumber,
-        });
-        _cancelTake(pillSheet, store);
-      },
-    );
-  }
-
-  Future<void> _take(
-    BuildContext context,
-    PillSheet pillSheet,
-    DateTime takenDate,
-    RecordPageStore store,
-  ) async {
-    if (pillSheet.todayPillNumber == pillSheet.lastTakenPillNumber) {
-      return;
-    }
-    try {
-      await store.take(takenDate);
-      _requestInAppReview();
-      Future.delayed(Duration(milliseconds: 500)).then((_) {
-        showReleaseNotePreDialog(context);
-      });
-    } catch (exception, stack) {
-      errorLogger.recordError(exception, stack);
-      store.handleException(exception);
-    }
-  }
-
-  void _cancelTake(PillSheet pillSheet, RecordPageStore store) {
-    if (pillSheet.todayPillNumber != pillSheet.lastTakenPillNumber) {
-      return;
-    }
-    final lastTakenDate = pillSheet.lastTakenDate;
-    if (lastTakenDate == null) {
-      return;
-    }
-    store.take(lastTakenDate.subtract(Duration(days: 1)));
   }
 
   PillSheetView _pillSheet(
@@ -281,7 +206,7 @@ class RecordPage extends HookWidget {
           return;
         }
         var takenDate = now().subtract(Duration(days: diff));
-        _take(context, pillSheet, takenDate, store);
+        take(context, pillSheet, takenDate, store);
       },
       premiumMarkBuilder: state.isPremium &&
               state.appearanceMode == PillSheetAppearanceMode.date
@@ -351,23 +276,5 @@ class RecordPage extends HookWidget {
         }
       },
     );
-  }
-
-  _requestInAppReview() {
-    SharedPreferences.getInstance().then((store) async {
-      final key = IntKey.totalCountOfActionForTakenPill;
-      int? value = store.getInt(key);
-      if (value == null) {
-        value = 0;
-      }
-      value += 1;
-      store.setInt(key, value);
-      if (value % 7 != 0) {
-        return;
-      }
-      if (await InAppReview.instance.isAvailable()) {
-        await InAppReview.instance.requestReview();
-      }
-    });
   }
 }
